Change script name and remove the export command
The export command is removed for the moment as the file it produced
was not really useful. Indeed, I choose the export format to be the
same as the format used by requirements files. The goal was to import
a mirror using the download command and the '-r' option with the
exported file. Unfortunately it doesn't work as expected when multiple
versions of the same package is present in the download directory as
the same package appears multiple times in the exported file and pip
doesn't like that (which is right).